ocrnet/requirements.txt
certifi==2017.11.5
chardet==3.0.4
cheroot==5.10.0
CherryPy==12.0.1
idna==2.6
jaraco.classes==1.4.3
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
portend==2.2
pyocr==0.4.7
pyPdf==1.13
pytesseract==0.1.7
pytz==2017.3
requests==2.18.4
six==1.11.0
tempora==1.9
tesserocr==2.2.2
urllib3==1.22
Wand==0.4.4
